Updated code structure
integrated code from dev-sathvika and added router to actuation page
