+++ package.json
@@ -16,8 +16,9 @@
     "react-icons": "^4.12.0",
     "react-input-slider": "^6.0.1",
     "react-leaflet": "^4.2.1",
-    "react-scripts": "5.0.1",
-    "sweetalert2": "^11.10.1",
+    "react-router-dom": "^6.21.2",
+    "react-scripts": "^5.0.1",
+    "sweetalert2": "^11.6.13",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -28,8 +29,7 @@
   },
   "eslintConfig": {
     "extends": [
-      "react-app",
-      "react-app/jest"
+      "react-app"
     ]
   },
   "browserslist": {
