+++ src/ActuationPage.js
@@ -1,0 +1,12 @@
+// ActuationPage.js
+import React from 'react';
+
+function ActuationPage() {
+  return (
+    <div>
+      <h1>Actuation Page</h1>
+    </div>
+  );
+}
+
+export default ActuationPage;
